Get rid of theme, use colour escape codes instead
After doing some research it seemed like it was going to be way too
complicated to do this with a theme.

I was able to print the message, but would need to include the whole
rest of the line, including timestamp, nick, etc. in order to make the
line look like it was a normal message. Too much work for not enough
gain. That's not to say it isn't possible. Here's the one script I found
that appeared to be doing this:
https://github.com/irssi/scripts.irssi.org/blob/d62bb05a34ffd6d8d0f719c551d8e454880ef8e1/scripts/nm2.pl

Crazy talented stuff.

Since we really just want to turn the message text grey, use a simple
`signal_continue` and colourise the text with escape codes.